refactor: Data,Encodable Extension 파일 분리
