+++ OpenMarket/OpenMarket/Network/APIProvider/APIProvider.swift
@@ -40,9 +40,7 @@ final class APIProvider<T: Decodable>: Provider {
             completion(.failure(error))
         }
     }
-}
-
-extension APIProvider {
+    
     private func checkError(
         with data: Data?,
         _ response: URLResponse?,
@@ -70,16 +68,5 @@ extension APIProvider {
         }
 
         completion(.success((data)))
-    }
-}
-
-fileprivate extension Data {
-    func decode<T: Decodable>() -> Result<T, Error> {
-        do {
-            let decoded = try JSONDecoder().decode(T.self, from: self)
-            return .success(decoded)
-        } catch {
-            return .failure(NetworkError.decodeError)
-        }
     }
 }
