+++ OpenMarket/OpenMarket/Network/EndPoint/Requestable.swift
@@ -76,17 +76,3 @@ extension Requestable {
         return .success(url)
     }
 }
-
-fileprivate extension Encodable {
-    func toDictionary() -> Result<[String: Any], NetworkError> {
-        do {
-            let data = try JSONEncoder().encode(self)
-            guard let jsonData = try JSONSerialization.jsonObject(with: data) as? [String: Any] else {
-                return .failure(.decodeError)
-            }
-            return .success(jsonData)
-        } catch {
-            return .failure(.decodeError)
-        }
-    }
-}
