+++ OpenMarket/OpenMarket/View/MainView.swift
@@ -15,5 +15,6 @@ final class MainView: UIView {
     
     required init?(coder: NSCoder) {
         super.init(coder: coder)
+        backgroundColor = .white
     }
 }
